Added in more Sprites and Animations
Added in the full sprite sheet for Gundham Tanaka with making the walking and Idle animation which is activated by code. Then added Monokuma as an enemy and all it's sprites from the sprite sheet. As well as making animations for it. Then had the physics material for the walls so the player doesn't get stuck on the wall as they play through the game.